Move getBitmapDescriptor into different file
